requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
Django==5.2.1
djangorestframework==3.16.0
kombu==5.5.3
mysql-connector==2.2.9
mysqlclient==2.2.7
paramiko==3.5.1
prompt_toolkit==3.0.51
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.0.0
six==1.17.0
sqlparse==0.5.3
sshtunnel==0.4.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
